Move all config files to src folder
The configs are technically source code in the plugin, so it makes more sense to have the rule declarations in src, while the configs at root bring it in from dist to format/lint itself. This feels cleaner than having an eslint.config.ts at the root alongside an eslint.self.config.js purely for self-linting.
