+++ src/configs/prettier-rules.ts
@@ -1,0 +1,7 @@
+import type { Config } from "prettier";
+
+const prettierRules: Config = {
+  printWidth: 100,
+};
+
+export default prettierRules;
